requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pyyaml>=6.0
google-generativeai>=0.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
python-dotenv>=1.0.0
